Use bezier curves to draw the SR logo
To get the logo to print nicely (using gaf export) you have to set the
line colour to the 'output-background' colour. This doesn't show up in
gschem though, so I've made two copies of the line that makes up the SR
part of the logo. One of them is shown in gschem, the other is shown in
print.